Share tests proper
